+++ src/options.rs
@@ -69,6 +69,10 @@ pub struct CMAESOptions {
     /// The value to use for the [`TerminationReason::TolXUp`][crate::TerminationReason::TolXUp]
     /// termination criterion. Default value is `1e+8`.
     pub tol_x_up: f64,
+    /// The value to use for the
+    /// [`TerminationReason::TolConditionCov`][crate::TerminationReason::TolConditionCov]
+    /// termination criterion. Default value is `1e+14`.
+    pub tol_condition_cov: f64,
     /// The seed for the RNG used in the algorithm. Can be set manually for deterministic runs. By
     /// default a random seed is used if this field is `None`.
     pub seed: Option<u64>,
@@ -98,6 +102,7 @@ impl CMAESOptions {
             tol_fun: 1e-12,
             tol_x: None,
             tol_x_up: 1e8,
+            tol_condition_cov: 1e14,
             seed: None,
             plot_options: None,
             print_gap_evals: None,
@@ -175,6 +180,13 @@ impl CMAESOptions {
     /// [`TerminationReason::TolXUp`][crate::TerminationReason::TolXUp]).
     pub fn tol_x_up(mut self, tol_x_up: f64) -> Self {
         self.tol_x_up = tol_x_up;
+        self
+    }
+
+    /// Changes the value for the `TolConditionCov` termination criterion from the default value
+    /// (see [`TerminationReason::TolConditionCov`][crate::TerminationReason::TolConditionCov]).
+    pub fn tol_condition_cov(mut self, tol_condition_cov: f64) -> Self {
+        self.tol_condition_cov = tol_condition_cov;
         self
     }