+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fixed `TolFun` termination criterion checking whether the current generation's function values are all below `tol_fun` instead of whether the range of them is.
 - Added `FunTarget` termination criterion that checks whether a function value threshold has been reached.
 - Added `tol_x_up` option to configure `TolXUp` termination criterion.
+- Renamed `ConditionCov` termination criterion to `TolConditionCov` and added `tol_condition_cov` option to configure it.
 
 ## 0.1.1 (March 4th, 2022)